requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.2.1
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
QDarkStyle==3.2.3
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtPy==2.4.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
shapely==2.1.0
six==1.17.0
sympy==1.13.1
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.4.0
